# sql: inline `addActiveQuery` anonymous function