--- a/pkg/sql/conn_executor_exec.go
+++ b/pkg/sql/conn_executor_exec.go
@@ -408,21 +408,15 @@ func (ex *connExecutor) execStmtInOpenState(
 	var queryDoneAfterFunc chan struct{}
 	var txnDoneAfterFunc chan struct{}
 
-	addActiveQuery := func() {
-		ctx, vars.cancelQuery = ctxlog.WithCancel(ctx)
-		ex.incrementStartedStmtCounter(vars.ast)
-		func(st *txnState) {
-			st.mu.Lock()
-			defer st.mu.Unlock()
-			st.mu.stmtCount++
-		}(&ex.state)
-		ex.addActiveQuery(parserStmt, pinfo, queryID, vars.cancelQuery)
-	}
-
 	// For pausable portal, the active query needs to be set up only when
 	// the portal is executed for the first time.
 	if !portal.isPausable() || !portal.pauseInfo.execStmtInOpenState.cleanup.isComplete {
-		addActiveQuery()
+		ctx, vars.cancelQuery = ctxlog.WithCancel(ctx)
+		ex.incrementStartedStmtCounter(vars.ast)
+		ex.state.mu.Lock()
+		ex.state.mu.stmtCount++
+		ex.state.mu.Unlock()
+		ex.addActiveQuery(parserStmt, pinfo, queryID, vars.cancelQuery)
 		if portal.isPausable() {
 			portal.pauseInfo.execStmtInOpenState.cancelQueryFunc = vars.cancelQuery
 			portal.pauseInfo.execStmtInOpenState.cancelQueryCtx = ctx