@mgartner 2024-12-04
`BenchmarkEndToEnd` shows the elimination of 1 allocation per query:

```
name                                     old time/op    new time/op    delta
EndToEnd/kv-read/vectorize=on/Simple        129µs ± 4%     132µs ± 5%  +1.95%  (p=0.008 n=18+17)
EndToEnd/kv-read/vectorize=on/Prepared     77.8µs ± 5%    80.6µs ± 8%  +3.58%  (p=0.004 n=20+19)
EndToEnd/kv-read/vectorize=off/Simple       157µs ± 3%     157µs ± 2%    ~     (p=0.753 n=18+19)
EndToEnd/kv-read/vectorize=off/Prepared    85.8µs ±10%    83.8µs ± 6%    ~     (p=0.070 n=19+20)

name                                     old alloc/op   new alloc/op   delta
EndToEnd/kv-read/vectorize=on/Simple       31.6kB ± 5%    31.4kB ± 1%    ~     (p=0.162 n=18+16)
EndToEnd/kv-read/vectorize=on/Prepared     20.3kB ± 0%    20.2kB ± 0%    ~     (p=0.066 n=16+16)
EndToEnd/kv-read/vectorize=off/Simple      32.5kB ± 0%    32.5kB ± 0%    ~     (p=0.845 n=17+17)
EndToEnd/kv-read/vectorize=off/Prepared    21.4kB ± 0%    21.4kB ± 0%  -0.11%  (p=0.007 n=17+17)

name                                     old allocs/op  new allocs/op  delta
EndToEnd/kv-read/vectorize=on/Simple          296 ± 0%       295 ± 1%  -0.38%  (p=0.000 n=13+16)
EndToEnd/kv-read/vectorize=on/Prepared        208 ± 0%       207 ± 0%  -0.49%  (p=0.000 n=18+19)
EndToEnd/kv-read/vectorize=off/Simple         288 ± 0%       287 ± 0%  -0.35%  (p=0.000 n=15+17)
EndToEnd/kv-read/vectorize=off/Prepared       200 ± 0%       199 ± 0%  -0.54%  (p=0.000 n=17+18)```