@RaduBerinde 2024-12-04
*[`pkg/sql/conn_executor_exec.go` line 422 at r2](https://reviewable.io/reviews/cockroachdb/cockroach/136730#-ODI7McKB9qyJn7KguTv:-ODITHIWDaC7zmiagKb4:ba9zd8b) ([raw file](https://github.com/cockroachdb/cockroach/blob/82dfebb289a6cda87423b811d2bcaf5eea639312/pkg/sql/conn_executor_exec.go#L422)):*
<details><summary><i>Previously, yuzefovich (Yahor Yuzefovich) wrote…</i></summary><blockquote>

Why did this change? Should we remove the argument to anonymous function then?
</blockquote></details>

Do we need this func at all? Can't we just Lock, increment, Unlock. It's not like we expect that the increment might panic


<!-- Sent from Reviewable.io -->